refactored flake8 happy
